Move golang out of asm package
