+++ internal/asm/assembler.go
@@ -95,10 +95,10 @@ type AssemblerBase interface {
 	CompileReadInstructionAddress(destinationRegister Register, beforeAcquisitionTargetInstruction Instruction)
 }
 
-// binaryOffsetMaximum represents the limit on the size of jump table in bytes.
+// JumpTableMaximumOffset represents the limit on the size of jump table in bytes.
 // When users try loading an extremely large webassembly binary which contains a br_table
 // statement with approximately 4294967296 (2^32) targets. Realistically speaking, that kind of binary
 // could result in more than ten giga bytes of native JITed code where we have to care about
 // huge stacks whose height might exceed 32-bit range, and such huge stack doesn't work with the
 // current implementation.
-const jumpTableMaximumOffset = math.MaxUint32
+const JumpTableMaximumOffset = math.MaxUint32
